fixed failing test cases

+++ src/com/redhat/qe/katello/tests/cli/OrgTests.java
@@ -266,6 +266,7 @@ public class OrgTests extends KatelloCliTestScript{
 		String org_name = "Raleigh-" + uniqueID;
 		String diff_org_name = "Durham-" + uniqueID;
 		KatelloOrg org = new KatelloOrg(org_name,null);
+		KatelloOrg diff_org = null;
 		exec_result = org.cli_create();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
 		KatelloProvider provider = new KatelloProvider(KatelloProvider.PROVIDER_REDHAT,org_name,null,null);
@@ -276,19 +277,24 @@ public class OrgTests extends KatelloCliTestScript{
 				System.getProperty("katello.server.sshkey.passphrase", "null"));
 		Assert.assertTrue(scp.sendFile("data"+File.separator+"stack-manifest.zip", "/tmp"),
 				"stack-manifest.zip sent successfully");			
+		try {
 		exec_result = provider.import_manifest("/tmp"+File.separator+"stack-manifest.zip", new Boolean(true));
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
-		KatelloOrg diff_org = new KatelloOrg(diff_org_name,null);
+		diff_org = new KatelloOrg(diff_org_name,null);
 		exec_result = diff_org.cli_create();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
 		provider = new KatelloProvider(KatelloProvider.PROVIDER_REDHAT,diff_org_name,null,null);
 		exec_result = provider.import_manifest("/tmp"+File.separator+"stack-manifest.zip", new Boolean(true));
-		Assert.assertTrue(exec_result.getExitCode() == 166, "Check - return code");
-		Assert.assertTrue(getOutput(exec_result).contains(String.format("This distributor has already been imported by another owner")),"Check - return string");
+		Assert.assertTrue(exec_result.getExitCode().intValue() == 144, "Check - return code");
+		Assert.assertTrue(getOutput(exec_result).contains("This subscription management application has already been imported by another owner."),"Check - return string");
+		} finally {
 		exec_result  = org.delete();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
-		exec_result = diff_org.delete();
-		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");	
+		if (diff_org != null) {
+			exec_result = diff_org.delete();
+			Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
+		}
+		}
 	}
 
  	
@@ -308,25 +314,29 @@ public class OrgTests extends KatelloCliTestScript{
 				System.getProperty("katello.server.sshkey.private", ".ssh/id_hudson_dsa"), 
 				System.getProperty("katello.server.sshkey.passphrase", "null"));
 		Assert.assertTrue(scp.sendFile("data"+File.separator+"stack-manifest.zip", "/tmp"),
-				"stack-manifest.zip sent successfully");			
+				"stack-manifest.zip sent successfully");	
+		try {
 		exec_result = provider.import_manifest("/tmp"+File.separator+"stack-manifest.zip", new Boolean(true));
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
 		Assert.assertTrue(getOutput(exec_result).contains(String.format("Manifest imported")),"Check - return string");
 		exec_result = provider.import_manifest("/tmp"+File.separator+"stack-manifest.zip", new Boolean(true));
-		Assert.assertTrue(exec_result.getExitCode() == 166, "Check - return code");
+		Assert.assertTrue(exec_result.getExitCode().intValue() == 144, "Check - return code");
 		Assert.assertTrue(getOutput(exec_result).contains(String.format("Import is the same as existing data")),"Check - return string");
+		} finally {
 		exec_result  = org.delete();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
+		}
 	}
 	
 	
-	@Test(description = "Delete a manifest from an ORG and upload the same to an other ORG",groups={"cfse-cli","headpin-cli"})
+	@Test(description = "Delete a manifest from an ORG and upload the same to an other ORG",groups={"headpin-cli", "cfse-ignore"})
 	public void test_ReUploadManifestDiffOrg(){
 		
 		String uniqueID = KatelloUtils.getUniqueID();
 		String org_name = "Raleigh-" + uniqueID;
 		String diff_org_name = "Durham-" + uniqueID;
 		KatelloOrg org = new KatelloOrg(org_name,null);
+		KatelloOrg diff_org = null;
 		exec_result = org.cli_create();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
 		KatelloProvider provider = new KatelloProvider(KatelloProvider.PROVIDER_REDHAT,org_name,null,null);
@@ -337,20 +347,25 @@ public class OrgTests extends KatelloCliTestScript{
 				System.getProperty("katello.server.sshkey.passphrase", "null"));
 		Assert.assertTrue(scp.sendFile("data"+File.separator+"stack-manifest.zip", "/tmp"),
 				"stack-manifest.zip sent successfully");			
+		try {
 		exec_result = provider.import_manifest("/tmp"+File.separator+"stack-manifest.zip", new Boolean(true));
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
 		exec_result = provider.delete_manifest();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
-		KatelloOrg diff_org = new KatelloOrg(diff_org_name,null);
+		diff_org = new KatelloOrg(diff_org_name,null);
 		exec_result = diff_org.cli_create();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
 		provider = new KatelloProvider(KatelloProvider.PROVIDER_REDHAT,diff_org_name,null,null);
 		exec_result = provider.import_manifest("/tmp"+File.separator+"stack-manifest.zip", new Boolean(true));
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
+		} finally {
 		exec_result	= org.delete();
 		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
-		exec_result = diff_org.delete();
-		Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
+		if (diff_org != null) {
+			exec_result = diff_org.delete();
+			Assert.assertTrue(exec_result.getExitCode() == 0, "Check - return code");
+		}
+		}
 
 	}
